Don't re-trigger Java validation for all opened Java files when a Java file is saved.
