+++ microprofile.jdt/org.eclipse.lsp4mp.jdt.core/src/main/java/org/eclipse/lsp4mp/jdt/core/project/AbstractConfigSource.java
@@ -218,7 +218,8 @@ public abstract class AbstractConfigSource<T> implements IConfigSource {
 		return null;
 	}
 
-	private void reset() {
+	@Override
+	public void reset() {
 		config = null;
 		propertyInformations = null;
 	}
