+++ microprofile.jdt/org.eclipse.lsp4mp.jdt.core/src/main/java/org/eclipse/lsp4mp/jdt/core/project/IConfigSource.java
@@ -100,4 +100,5 @@ public interface IConfigSource {
 	 */
 	Set<String> getAllKeys();
 
+	void reset();
 }
